==> code_process_data_Python/code/00_c_translate_arabic_to_english.ipynb <==
# ! pip install transformers sentencepiece

# Code to translate document from arabic to english

from openpyxl import load_workbook

from ipywidgets import FloatProgress, VBox
from IPython.display import display

# Create a progress bar
progress_bar = FloatProgress(value=0, min=0, max=100, description='Loading:')

# Styling through layout (instead of .style)
progress_bar.style = {'description_width': 'initial'}  # Update style directly
progress_bar.layout.width = '500px'  # Adjust the width

# Display progress bar
display(progress_bar)

# Update progress
for i in range(101):
    progress_bar.value = i

import pandas as pd
from transformers import MarianMTModel, MarianTokenizer

# Load MarianMT model for Arabic-to-English
model_name = 'Helsinki-NLP/opus-mt-ar-en'
tokenizer = MarianTokenizer.from_pretrained(model_name)
model = MarianMTModel.from_pretrained(model_name)

# Define a function to translate text
def translate_text(text):
    if pd.isna(text):  # Handle NaN values
        return text
    try:
        inputs = tokenizer(text, return_tensors="pt", padding=True)
        outputs = model.generate(**inputs)
        return tokenizer.decode(outputs[0], skip_special_tokens=True)
    except Exception as e:
        print(f"Error translating {text}: {e}")
        return text


# Load the Excel file
file_path = '../../1. Data available/ISHTM_Hadhreen/Oct. Beneficiary Survey/Completed_Survey Oct 2024.xlsx'
data = pd.read_excel(file_path)

from openpyxl import load_workbook

def get_sheetnames_xlsx(filepath):
    wb = load_workbook(filepath, read_only=True, keep_links=False)
    return wb.sheetnames

sheet_name = get_sheetnames_xlsx(file_path)
sheet_name_translate = [translate_text(name) for name in sheet_name]

## Open the where we want to save the output
path = '../output/translated_beneficiary_survey_oct_2024.xlsx'
book = load_workbook(path)
writer = pd.ExcelWriter(path, engine = 'openpyxl')

for index, name in enumerate(sheet_name):
    data = pd.read_excel(file_path, sheet_name = name)
    # Translate column names
    data.columns = [translate_text(col) for col in data.columns]
    # Apply translation to the desired columns
    columns_to_translate = data.columns  # Replace with your actual column names
    for col in columns_to_translate:
        data[col] = data[col].apply(translate_text)
    data.to_excel(writer, sheet_name = sheet_name_translate[index])
writer.close()
